Login added and styling
Login and signup functionality added using Devise and a user model.
Webapps navbar changes depending on login status.

Styling done with Bootstrap, basic navbar layout with the %yield%
rendered in the body. The navbar contents react to the users login
status.

Layout of the patients view improved, styled into a table with
bootstrap.